% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_none}
\alias{add_none}
\title{Add NULL to choices}
\usage{
add_none(choices)
}
\arguments{
\item{choices}{[\code{character()}]\cr
Choices to augment with NULL.}
}
\value{
A \code{character} vector with NULL option prepended to \code{choices}.
}
\description{
Adds string-encoded NULL labeled as "none" to a vector of
choices. This is due to NULL not being able to be an element of a vector.
}
