% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prettify_chains.R
\name{add_sequence_info}
\alias{add_sequence_info}
\title{Add sequence info}
\usage{
add_sequence_info(content, seq_length, name)
}
\arguments{
\item{content}{[\code{character(1)}]\cr
Sequence with indices and newlines.}

\item{seq_length}{[\code{integer(1)}]\cr
Length of the sequence.}

\item{name}{[\code{character(1)}]\cr
Chain name, may be \code{NA} if not applicable.}
}
\value{
A single string with all info prepended and separated with newlines.
}
\description{
Prepends the sequence and indices with sequence length and chain
name (if applicable).
}
