% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ag_colnames}
\alias{ag_colnames}
\title{Find AmyloGraph column names in data}
\usage{
ag_colnames(data)
}
\arguments{
\item{data}{[\code{data.frame()}]\cr
Data, which column names are to be filtered.}
}
\value{
A string vector with filtered column names.
}
\description{
Filters column names of a data frame or matrix which fulfill
a custom role in AmyloGraph, like "interactor name".
}
