% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ag_data_groups}
\alias{ag_data_groups}
\title{Build color data for question answers}
\usage{
ag_data_groups()
}
\value{
A list with two elements:
\itemize{
\item \code{data}: a list of tibbles for each question; each tibble containing two
columns -- \code{values} (containing answers to questions) and \code{colors} (with
HTML color codes, all unique within a tibble),
\item \code{groups}: a list with translations between human- and computer-friendly
names for questions.
}
}
\description{
Build color data for question answers
}
