% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ag_data_interactions}
\alias{ag_data_interactions}
\title{Load AmyloGraph interactions list}
\usage{
ag_data_interactions()
}
\value{
\code{data.frame} coming from \code{interactions_data.csv} file with the
following changes:
\itemize{
\item \code{from_id} and \code{to_id} are unique hashes of interactor and interactee names
\item \code{interactor_sequence} and \code{interactee_sequence} are transformed with
\code{\link{read_chains}()} to a vector of tibbles.
}
}
\description{
Load AmyloGraph interactions list
}
