% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ag_data_nodes}
\alias{ag_data_nodes}
\title{Build list of proteins from interactions list}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 3 columns.
}
\usage{
ag_data_nodes
}
\value{
\code{data.frame} with a list of proteins that interact with others,
described with \code{label} (human-readable name), \code{id} (machine-readable name),
and \code{shape} (for the purpose of drawing a graph only).
}
\description{
Build list of proteins from interactions list
}
\keyword{datasets}
