% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ag_data_proteins}
\alias{ag_data_proteins}
\title{Load AmyloGraph proteins list}
\usage{
ag_data_proteins()
}
\value{
\code{data.frame} coming from \code{protein_data.csv} file with the
following changes:
\itemize{
\item \code{id} contains unique hashes of protein names.
}
}
\description{
Load AmyloGraph proteins list
}
