% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{ag_data_references}
\alias{ag_data_references}
\title{AmyloGraph references list}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 173 rows and 6 columns.
}
\usage{
ag_data_references
}
\value{
\code{data.frame} coming from \code{reference_table.csv} file with no changes.
}
\description{
AmyloGraph references list
}
\keyword{datasets}
