% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ag_load_data}
\alias{ag_load_data}
\title{Load all AmyloGraph data}
\usage{
ag_load_data()
}
\value{
A list with four elements:
\itemize{
\item \code{interactions}: see \code{\link{ag_data_interactions}()}
\item \code{groups}: see \code{\link{ag_data_groups}()}
\item \code{nodes}: see \code{\link{ag_data_nodes}()}
\item \code{proteins}: see \code{\link{ag_data_proteins}()}
}
}
\description{
Load all AmyloGraph data
}
