% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ag_option}
\alias{ag_option}
\title{Extract AmyloGraph option}
\usage{
ag_option(option)
}
\arguments{
\item{option}{[\code{character(1)}]\cr
Option name without leading \code{"ag-"}.}
}
\value{
Option value.
}
\description{
This function accesses the current value of a selected
AmyloGraph option.
}
\details{
The currently available options are:
\itemize{
\item \code{str_null}
\item \code{palette}
\item \code{side_panel_width}
\item \code{main_panel_width}
\item \code{colnames}
\item \code{interaction_attrs}
\item \code{center_network}
\item \code{sequence_group_length}
\item \code{chain_separator}
}
}
