% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ag_references}
\alias{ag_references}
\title{Load AmyloGraph references list}
\usage{
ag_references()
}
\value{
\code{data.frame} coming from \code{reference_table.csv} file with no changes.
}
\description{
Load AmyloGraph references list
}
