% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGID_button.R
\name{AGID_button}
\alias{AGID_button}
\title{Generate AGID buttons}
\usage{
AGID_button(interaction_ids)
}
\arguments{
\item{interaction_ids}{[\code{character()}]\cr
AGIDs to use for buttons.}
}
\value{
A \code{character} vector the same length as \code{interaction_ids}, each
element being a string representation of a button.
}
\description{
Generates a vector of AGID buttons, where AGID codes are labels
of the buttons. Each button sets the \code{selected_interaction} value to the
respective AGID on click.
}
