% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{citify}
\alias{citify}
\title{Generate paper citation}
\usage{
citify(reference)
}
\arguments{
\item{reference}{[\code{character matrix(1, )}]\cr
Paper citation data with the following columns: \code{all_names}, \code{title},
\code{journal}, \code{year}, and \code{doi}.}
}
\value{
A single string with well-formatted citation.
}
\description{
This function formats paper reference data in a human-friendly
format, akin to one of those used in formatting citations.
}
\seealso{
\code{\link{linkify_doi}()}, \code{\link{linkify_uniprot}()}
}
