% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-buttons.R
\name{collect_tags}
\alias{collect_tags}
\title{Collect button tags}
\usage{
collect_tags(buttons)
}
\arguments{
\item{buttons}{[\code{ag_button()}]\cr
Buttons whose tags to collect.}
}
\value{
A vector of tags (a \code{character} vector).
}
\description{
This functions collects all unique tags assigned to supplied
buttons.
}
