% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_graph_data.R
\name{color_by_attribute}
\alias{color_by_attribute}
\title{Add colors depending on attribute value}
\usage{
color_by_attribute(data, group)
}
\arguments{
\item{data}{[\code{tibble()}]\cr
Data to modify.}

\item{group}{[\code{character(1)}]\cr
Name of a currently selected attribute (if none, empty string).}
}
\value{
A \code{tibble} with additional \code{color} column of character type.
}
\description{

}
