% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-motifs.R
\name{contains_motif}
\alias{contains_motif}
\title{Check whether sequences fit motif}
\usage{
contains_motif(sequences, motif)
}
\arguments{
\item{sequences}{[\code{list()}]\cr
Sequences to check, each element being a \code{tibble} with \code{name} and \code{sequence}
columns.}

\item{motif}{[\code{character(1)}]\cr
Motif to look for in sequences.}
}
\value{
A logical vector of the same length as \code{sequences} vector, each
element showing whether a motif was found in a sequence.
}
\description{
Compare each sequence against a motif. Motif is patternized
before making comparisons.
}
