% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-motifs.R
\name{correct_motif}
\alias{correct_motif}
\title{Check motif correctness}
\usage{
correct_motif(motif)
}
\arguments{
\item{motif}{[\code{character(1)}]\cr
Motif to check.}
}
\value{
A single logical value.
}
\description{
Validates a motif against a regex.
}
