% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_num_interactions.R
\name{count_unique_interactors}
\alias{count_unique_interactors}
\alias{count_unique_interactees}
\title{Count unique interacting proteins}
\usage{
count_unique_interactors(protein_id)

count_unique_interactees(protein_id)
}
\arguments{
\item{protein_id}{[\code{character(1)}]\cr
ID of protein interactions are counted for.}

\item{interactions}{[\code{data.frame()}]\cr
AmyloGraph interaction data.}
}
\value{
An \code{integer} count of distinct protein interacting.
}
\description{
Counts number of unique interactors/interactees in AmyloGraph
database for a given protein.
}
