% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_num_interactions.R
\name{count_interactions}
\alias{count_interactions}
\title{Count interactions for a protein}
\usage{
count_interactions(protein_id)
}
\arguments{
\item{protein_id}{[\code{character(1)}]\cr
ID of protein interactions are counted for.}

\item{interactions}{[\code{data.frame()}]\cr
AmyloGraph interaction data.}
}
\value{
An \code{integer} count of interactions.
}
\description{
Counts number of interactions in AmyloGraph database for a given
protein.
}
