% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-buttons.R
\name{define_button}
\alias{define_button}
\title{Define a button type}
\usage{
define_button(ui, server, tags = NULL)
}
\arguments{
\item{ui}{[\code{function()}]\cr
Mostly HTML part of a button that's generated during initialization of a
button instance. Must contain \code{ns} parameter.}

\item{server}{[\code{function()}]\cr
Server logic that handles button-related actions. Must contain \code{input}
and \code{output} parameters.}

\item{tags}{[\code{character()}]\cr
Categories to apply to the button type (defined in \code{BUTTON_TAGS}).}
}
\value{
An \code{ag_button} object.
}
\description{
Defines a button type with its visible component, server logic,
and tags.
}
