% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-read_chains.R
\name{deparse_chains}
\alias{deparse_chains}
\title{Revert chain tibbles to strings}
\usage{
deparse_chains(tbl_sq, separator = ag_option("chain_separator"))
}
\arguments{
\item{tbl_sq}{[\code{tibble()}]\cr
Sequence data for one protein.}

\item{separator}{[\code{character(1)}]\cr
A character or string to join lines of data with.}
}
\value{
A character vector with FASTA-like entries or plain sequences.
}
\description{
Deparses chain data tibbles back to strings, depending on the
structure of the tibble; plain sequence if there's only one row and \code{name}
column is \code{NA}, a FASTA-like format otherwise.
}
