% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-read_chains.R
\name{empty_buffer}
\alias{empty_buffer}
\title{Flush chain reader buffer}
\usage{
empty_buffer(reader)
}
\arguments{
\item{reader}{[\code{AG_sequence_reader(1)}]\cr
Sequence reader with buffer to empty.}
}
\value{
\code{AG_sequence_reader} with empty buffer and additional sequence data.
}
\description{
Move chain data (name and sequence) from buffer to result
storage.
}
