% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_base_data.R
\name{filter_by_attribute}
\alias{filter_by_attribute}
\title{Filter by an attribute}
\usage{
filter_by_attribute(data, values)
}
\arguments{
\item{data}{[\code{tibble()}]\cr
Data to modify.}

\item{values}{[\code{ag_attr_values()}]\cr
Attribute filter values to keep.}
}
\value{
A \code{tibble} with filtered observations.
}
\description{
Filters data by attribute answer.
}
