% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_base_data.R
\name{filter_by_motif}
\alias{filter_by_motif}
\title{Filter by motif}
\usage{
filter_by_motif(data, motif)
}
\arguments{
\item{data}{[\code{tibble()}]\cr
Data to modify.}

\item{motif}{[\code{ag_motif(1)}]\cr
Motif to filter on.}
}
\value{
A \code{tibble} with filtered observations.
}
\description{
Filters data on a motif. Either \code{interactor_sequence} or
\code{interactee_sequence} must contain given motif.
}
