% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-GoogleAnalytics.R
\name{GoogleAnalytics}
\alias{GoogleAnalytics}
\title{Add Google Analytics to shiny app}
\usage{
GoogleAnalytics()
}
\value{
A list of tags to place inside a shiny app.
}
\description{
Includes \verb{<head>} and \verb{<body>} tags with Google Analytics
cookies scripts. Adapted from:
https://github.com/DataScienceScotland/shiny_cookies.
}
