% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_graph_data.R
\name{group_by_attribute}
\alias{group_by_attribute}
\title{Group data by attribute value}
\usage{
group_by_attribute(data, group)
}
\arguments{
\item{data}{[\code{tibble()}]\cr
Data to modify.}

\item{group}{[\code{character(1)}]\cr
Name of a currently selected attribute (if none, empty string).}
}
\value{
A \code{tibble} where each observation corresponds to a single edge on
the graph.
}
\description{
Groups interaction data by start and end nodes, and attribute
value (if applicable). Generates edge labels from all DOIs for given group.
}
