% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_handler.R
\name{HTML_download_handler}
\alias{HTML_download_handler}
\title{Handle download of AmyloGraph network to HTML}
\usage{
HTML_download_handler(node_positions, nodes)
}
\arguments{
\item{node_positions}{[\code{reactive(data.frame())}]\cr
Table of node positions with \code{x}, \code{y}, and \code{id} columns.}

\item{nodes}{[\code{data.frame()}]\cr
AmyloGraph node data.}

\item{edges}{[\code{data.frame()}]\cr
AmyloGraph edge data.}
}
\value{
A \code{downloadHandler} object.
}
\description{
Creates a download handler that can be inserted into an app and
used to download AmyloGraph network data in HTML format.
}
