% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{if_null_else}
\alias{if_null_else}
\title{Execute code if not NULL}
\usage{
if_null_else(value, code = identity, ...)
}
\arguments{
\item{value}{[\code{any}]\cr
Value to compare to \code{NULL}.}

\item{code}{[\code{function(1)}]\cr
Function that takes \code{value} as first (unnamed) parameter, executed if
\code{value} is not \code{NULL}.}

\item{...}{\cr
Other parameters passed to \code{code()}.}
}
\value{
\code{NULL} if input is \code{NULL}, return value of \code{code()} otherwise.
}
\description{
Checks input value for equality to \code{NULL} and executes the code
only if this equality is not satisfied.
}
