% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prettify_chains.R
\name{index_line}
\alias{index_line}
\title{Create a line of indices}
\usage{
index_line(from, to, by)
}
\arguments{
\item{from}{[\code{integer(1)}]\cr
First element of a sequence.}

\item{to}{[\code{integer(1)}]\cr
Last element of a sequence.}

\item{by}{[\code{integer(1)}]\cr
Size of a group and padding length.}
}
\value{
A vector of strings, each element being a number left-padded with
spaces.
}
\description{
Creates a vector of left-padded strings with indices in the
indicated range.
}
