% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{invert_names}
\alias{invert_names}
\title{Swap values and names}
\usage{
invert_names(value)
}
\arguments{
\item{value}{[\code{vector()}]\cr
Named vector with elements that should be coercible to a character vector,
preferably strings.}
}
\value{
A character vector.
}
\description{
Swaps values and names of a vector.
}
