% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_node_selected}
\alias{is_node_selected}
\title{Check if node is selected}
\usage{
is_node_selected(id)
}
\arguments{
\item{id}{[\code{character(0) | character(1)}]\cr
Node ID to verify.}
}
\value{
A single logical value.
}
\description{
Checks whether \code{id} contains any data and whether this data is
not equal \code{str_null} AG option.
}
