% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{label_and_order}
\alias{label_and_order}
\title{Label and reorder a vector}
\usage{
label_and_order(value, label)
}
\arguments{
\item{value}{[\code{vector()}]\cr
Values to label and reorder.}

\item{label}{[\code{character()}]\cr
Names to use when labeling and reordering.}
}
\value{
Reordered \code{value} vector with names.
}
\description{
Sets names to a vector and then reorders it so that names are
sorted alphabetically.
}
