% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-linkify.R
\name{linkify_doi}
\alias{linkify_doi}
\title{Wrap DOI in a link}
\usage{
linkify_doi(doi, truncate = TRUE)
}
\arguments{
\item{doi}{[\code{character()}]\cr
DOI codes to use (as labels and links).}

\item{truncate}{[\code{logical(1)}]\cr
Whether displayed DOI code should be truncated to the first 18 characters.}
}
\value{
A string vector of the same length as \code{doi} parameter, each element
being an \verb{<a>} tag for respective DOI code.
}
\description{
Wraps DOI code in a HTML clickable link to the respective
address on doi.org. Allows displaying a truncated DOI string instead of the
full form.
}
\seealso{
\code{\link{citify}()}, \code{\link{linkify_uniprot}()}
}
