% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{linkify_uniprot}
\alias{linkify_uniprot}
\title{Wrap UniProt ID in a link}
\usage{
linkify_uniprot(uniprot_id)
}
\arguments{
\item{uniprot_id}{[\code{character()}]\cr
UniProt IDs to use (as labels and links).}
}
\value{
A string vector of the same length as \code{uniprot_id} parameter, each
element being an \verb{<a>} tag for respective UniProt ID.
}
\description{
Wraps UniProt ID code in a HTML clickable link to the respective
address on uniprot.org.
}
\seealso{
\code{\link{citify}()}, \code{\link{linkify_doi}()}
}
