% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_js_code}
\alias{load_js_code}
\title{Read JavaScript code from file}
\usage{
load_js_code(name)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
File name without .js extension.}
}
\value{
A single string with JavaScript code extracted from the file.
}
\description{
Reads JS code from \code{inst/js} folder inside AmyloGraph package.
}
