% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_texts.R
\name{markdown_ag_version}
\alias{markdown_ag_version}
\title{Format AmyloGraph version into human-friendly string}
\usage{
markdown_ag_version()
}
\value{
A single string.
}
\description{
This function retrieves version number of currently installed
\pkg{AmyloGraph} package and formats it into a neatly looking sentence.
}
