% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_row_selection.R
\name{observe_deselect_button}
\alias{observe_deselect_button}
\title{Observe "Deselect" button}
\usage{
observe_deselect_button(ns, button_id, any_row_selected)
}
\arguments{
\item{ns}{[\code{function(1)}]\cr
Namespace-generating function that takes single string as the only argument.}

\item{button_id}{[\code{character(1)}]\cr
ID of the "Deselect" button to observe.}

\item{any_row_selected}{[\code{reactive(logical(1))}]\cr
A reactive value telling whether any row is selected in a table.}
}
\description{
Toggles button on and off based on whether any row is selected
in a table.
}
