% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_selecting_all.R
\name{observe_deselecting_all}
\alias{observe_deselecting_all}
\title{Remove selections in a table}
\usage{
observe_deselecting_all(input, button_id, table_proxy)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the table.}

\item{button_id}{[\code{character(1)}]\cr
ID of a button that triggers removing selections.}

\item{table_proxy}{[\code{dataTableProxy(1)}]\cr
Proxy of a table to update.}
}
\description{
Sets selected rows to none in a table in response to clicking
a button.
}
