% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_edges_change.R
\name{observe_edges_change}
\alias{observe_edges_change}
\title{Observes graph edges changing}
\usage{
observe_edges_change(input, edges)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the graph.}

\item{edges}{[\code{reactivevalues()}]\cr
AmyloGraph data with "graph" element.}
}
\description{
Observes graph edges changing
}
