% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_filter_active.R
\name{observe_filter_active}
\alias{observe_filter_active}
\title{Toggle active state of a filter checkbox}
\usage{
observe_filter_active(group, attribute)
}
\arguments{
\item{group}{[\code{reactive(character(1))}]\cr
Name of a currently selected attribute (if none, empty string).}

\item{attribute}{[\code{character(1)}]\cr
One of interaction attributes, relevant for observed filter.}
}
\description{
Sets a filter in an active state if and only if the currently
selected group is equal to filter's attribute.
}
