% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_interaction_selection.R
\name{observe_interaction_selection}
\alias{observe_interaction_selection}
\title{Display "Single interaction" tab on selection}
\usage{
observe_interaction_selection(input)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app.}
}
\description{
Updates tabset panel to show "Single interaction" tab on the
tab list and moves the user there if an interaction is selected. If an
interaction is deselected, hides the tab and moves the user back to graph.
}
