% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_moving_selection.R
\name{observe_moving_selection}
\alias{observe_moving_selection}
\title{Copy selections in subtables to main table}
\usage{
observe_moving_selection(
  input,
  subtables,
  edges,
  table_proxy,
  tabs_visited,
  initial_selection
)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app.}

\item{subtables}{[code{reactivevalues()}]\cr
A list with two subtables for interactions as interactor and interactee
respectively.}

\item{edges}{[\code{reactivevalues()}]\cr
AmyloGraph data with "table" element.}

\item{table_proxy}{[\code{dataTableProxy(1)}]\cr
Proxy of the main table (target table).}

\item{rvals}{[\code{reactivevalues}]\cr
Reactive values handling the problem of setting selection in a table that
wasn't displayed previously.}
}
\description{
On "select" button click collects all selected interactions in
both subtables in single protein panel. Then finds these interactions in the
main table and sets the selected rows to these. Lastly, moves the user to
the table tab.
}
