% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_node_selection.R
\name{observe_node_selection}
\alias{observe_node_selection}
\title{Display single protein panel on selection}
\usage{
observe_node_selection(input)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app.}
}
\value{
An observer.
}
\description{
Displays the panel with details of a single protein when
a protein (a node) is selected and hides when not. Moreover, it handles
the changes to selected node of a graph.
}
