% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-motifs.R
\name{patternize_motif}
\alias{patternize_motif}
\title{Upgrade motif to detect ambiguous letters}
\usage{
patternize_motif(motif)
}
\arguments{
\item{motif}{[\code{character(1)}]\cr
Motif to patternize.}
}
\value{
A single string with regex-ready motif.
}
\description{
Replaces ambiguous letters (\code{B}, \code{J}, \code{Z}, \code{X}, and \code{*}) with
a set of characters (e.g. \code{B} with \verb{[BDN]}).
}
