% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pluralize}
\alias{pluralize}
\title{Return plural or singular form}
\usage{
pluralize(value, plural = "s", singular = "")
}
\arguments{
\item{value}{[\code{numeric()}]\cr
Values to check for equality to 1.}

\item{plural}{[\code{character(1)}]\cr
String to return if equality not satisfied.}

\item{singular}{[\code{character(1)}]\cr
String to return if equality satisfied.}
}
\value{
A vector of plural/singular values of the same length as \code{value}
parameter.
}
\description{
Returns a plural or singular form depending on whether \code{value}
is equal 1. By default it returns plural suffix "-s" or singular empty
suffix.
}
