% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prettify_chains.R
\name{prettify_chains}
\alias{prettify_chains}
\title{Present chains in human-friendly format}
\usage{
prettify_chains(tbl_sq)
}
\arguments{
\item{tbl_sq}{[\code{data.frame()}]\cr
Sequence data for one protein, a result of a call to
\code{\link{read_chains}()}.}
}
\value{
A single string, possibly with multiple lines.
}
\description{
Converts a tibble with sequence data for chains of a protein to
a human-readable string. If tibble is empty, returns \code{"no sequence available"}
instead.
}
