% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prettify_chains.R
\name{prettify_sequence_output}
\alias{prettify_sequence_output}
\title{Present sequence in human-friendly format}
\usage{
prettify_sequence_output(name, sequence)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Chain name. If not applicable, use \code{NA}.}

\item{sequence}{[\code{character(1)}]\cr
Amino acid sequence.}
}
\value{
A single string with multiple lines (probably 3 or 4).
}
\description{
Displays a sequence in a line in group of 10 (by default,
configurable with \code{sequence_group_length} AG option), each group numbered.
Prepends it with sequence length and chain name, if applicable.
}
