% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_allow_selection_transfer.R
\name{reactive_allow_selection_transfer}
\alias{reactive_allow_selection_transfer}
\title{Verify if "transfer selection" allowed}
\usage{
reactive_allow_selection_transfer(input)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the protein details.}
}
\value{
A \code{reactive} object with a single boolean describing whether
transfering selection is allowed.
}
\description{
This function checks if it is allowed to transfer selection from
protein interaction tables to the main table. The requirements are that there
are rows selected in either interactors or interactees table, and "ignore
filters" option is not used.
}
