% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_base_data.R
\name{reactive_base_data}
\alias{reactive_base_data}
\title{Generate base tabular data}
\usage{
reactive_base_data(motif, ...)
}
\arguments{
\item{motif}{[\code{reactive(ag_motif(1))}]\cr
Motif to filter on.}

\item{...}{[\code{reactive(ag_attr_values())}]\cr
Attribute value filters to apply.}
}
\value{
A \code{reactive} object with a \code{tibble} containing filtered
base tabular data.
}
\description{
Applies a set of filters to \code{ag_data_interactions} based on
a set of attribute value filters and motif filter.
}
