% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_graph_data.R
\name{reactive_graph_data}
\alias{reactive_graph_data}
\title{Generate graph data}
\usage{
reactive_graph_data(data, group)
}
\arguments{
\item{data}{[\code{reactive(tibble())}]\cr
AmyloGraph tabular edge data.}

\item{group}{[\code{reactive(character(1))}]\cr
Name of a currently selected attribute (if none, empty string).}
}
\value{
A \code{reactive} object with a \code{tibble} containing mutated
graph data.
}
\description{
Applies a set of transformations to base tabular data based on
currently selected attribute to group by.
}
