% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_node_positions.R
\name{reactive_node_positions}
\alias{reactive_node_positions}
\title{Retrieve node positions in a network}
\usage{
reactive_node_positions(input, network_id)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the network.}

\item{network_id}{[\code{character(1)}]\cr
ID of the network.}
}
\value{
A \code{reactive} object with \code{NULL} if no positions available or
a \code{tibble} with \code{x}, \code{y} and \code{id} columns.
}
\description{
Triggers update of \code{network_positions} input value, then reads
this value and formats it as a \code{data.frame}.
}
