% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_selected_node.R
\name{reactive_selected_node_info}
\alias{reactive_selected_node_info}
\title{Access info of currently selected protein}
\usage{
reactive_selected_node_info(input, node_data)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the protein selector.}

\item{node_data}{[\code{data.frame()}]\cr
AmyloGraph node data.}
}
\value{
A \code{reactive} object with a \code{data.frame} with exactly one
row of data, corresponding to currently selected node/protein.
}
\description{
This function accesses information about protein selected in
one of the available selectors (drop-down list or clicking on graph node).
}
