% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_selected_node.R
\name{reactive_selected_node_label}
\alias{reactive_selected_node_label}
\title{Access label of currently selected protein}
\usage{
reactive_selected_node_label(selected_node_info)
}
\arguments{
\item{selected_node_info}{[\code{reactive(data.frame())}]\cr
A row from AmyloGraph node data for a single protein.}
}
\value{
A \code{reactive} object with a single string containing protein
name.
}
\description{
This function extracts label from protein info.
}
