% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_selection_config.R
\name{reactive_selection_config}
\alias{reactive_selection_config}
\title{Construct DT selection config}
\usage{
reactive_selection_config(rvals)
}
\arguments{
\item{rvals}{[\code{reactivevalues()}]\cr
Information about table having been visited and rows selected in single
protein tables.}
}
\value{
Value to use as \code{selection} parameter in \code{renderDataTable()}, either
a list or a string.
}
\description{
Constructs selection configuration for the main table according
to DataTable specification. It's required due to table proxy not being
available before rendering the table first.
}
