% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_table_data.R
\name{reactive_subtable_data}
\alias{reactive_subtable_data}
\title{Build table data for single protein}
\usage{
reactive_subtable_data(edges, ns, input, target_id, target_variable)
}
\arguments{
\item{edges}{[\code{reactivevalues()}]\cr
AmyloGraph data with "table" element.}

\item{ns}{[code{function(1)}]\cr
Namespace function to use for AGID buttons.}

\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the protein selector.}

\item{target_id}{[\code{character(1)}]\cr
ID column name to use.}

\item{target_variable}{[\code{character(1)}]\cr
Protein name column to use.}
}
\value{
A \code{reactive} object with a \code{tibble} containing filtered
interaction data for a single protein, either with interactors or
interactees (depends on selected target ID).
}
\description{
Mutates filtered interaction data for a single selected protein
to contain DOI links and clickable AGIDs; then selects only columns that are
displayed in the interactors or interactees table.
}
