% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_table_data.R
\name{reactive_table_data}
\alias{reactive_table_data}
\title{Build table data}
\usage{
reactive_table_data(edges, ns)
}
\arguments{
\item{edges}{[\code{reactivevalues()}]\cr
AmyloGraph data with "table" element.}

\item{ns}{[code{function()}]\cr
Namespace function to use for AGID buttons.}
}
\value{
A \code{reactive} object with a \code{tibble} containing filtered
interaction data.
}
\description{
Mutates filtered interaction data to contain DOI links and
clickable AGIDs; then selects only columns that are displayed in the main
table.
}
