% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-read_chains.R
\name{read_chains}
\alias{read_chains}
\title{Parse chain sequence data}
\usage{
read_chains(txt, separator = ag_option("chain_separator"))
}
\arguments{
\item{txt}{[\code{character(1)}]\cr
Sequence data to parse.}

\item{separator}{[\code{character(1)}]\cr
A character or string to split data into lines on.}
}
\value{
A \code{tibble} with the following columns: \code{name} and \code{sequence}.
}
\description{
Reads amino acid sequences and their corresponding names, if
available. The input should be either in FASTA format or a single sequence
without a name.
}
