% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_num_interactions.R
\name{render_num_interactions_by_paper}
\alias{render_num_interactions_by_paper}
\title{Render a plot of interaction count by paper}
\usage{
render_num_interactions_by_paper(interactions, ...)
}
\arguments{
\item{interactions}{[\code{data.frame()}]\cr
AmyloGraph interaction data.}

\item{...}{\cr
Additional parameters to \code{renderPlot()}.}
}
\description{
Renders a barplot of the frequency of interaction counts for
papers; this shows how many interactions are usually retrieved from a single
paper.
}
