% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_num_interactions.R
\name{render_num_interactions_by_protein}
\alias{render_num_interactions_by_protein}
\title{Render a table of interaction count by protein}
\usage{
render_num_interactions_by_protein()
}
\arguments{
\item{interactions}{[\code{data.frame()}]\cr
AmyloGraph interaction data.}

\item{nodes}{[\code{data.frame()}]\cr
AmyloGraph node data.}
}
\description{
Renders a table with interaction count for each protein,
arranged by protein name alphabetically.
}
