% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_protein_info.R
\name{render_protein_info}
\alias{render_protein_info}
\title{Render list of protein sources}
\usage{
render_protein_info(protein_data, selected_node_label)
}
\arguments{
\item{protein_data}{[\code{data.frame}]\cr
A list of source organisms and their UniProt IDs for all proteins in the
database.}

\item{selected_node_label}{[\code{reactive(character(1))}]\cr
A label (i.e. human-readable name) of a protein currently selected on the
graph.}
}
\value{
A \code{renderText} object.
}
\description{
This function generates a list of source organisms with their
UniProt IDs for a given protein. UniProt IDs contain links to UniProt site.
}
