% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prettify_chains.R
\name{safe_seq}
\alias{safe_seq}
\title{Safely create a sequence of numbers}
\usage{
safe_seq(from, to, ...)
}
\arguments{
\item{from}{[\code{numeric(1)}]\cr
First element of a sequence.}

\item{to}{[\code{numeric(1)}]\cr
Upper limit of a sequence.}

\item{...}{\cr
Arguments passed to \code{\link[base]{seq}()}.}
}
\value{
A vector of numbers.
}
\description{
Wraps \code{base::seq()} with a safeguard that returns an empty
vector if sequence cannot be created.
}
