% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_graph_data.R
\name{select_graph_columns}
\alias{select_graph_columns}
\title{Select columns relevant to graph data}
\usage{
select_graph_columns(data, group)
}
\arguments{
\item{data}{[\code{tibble()}]\cr
Data to modify.}

\item{group}{[\code{character(1)}]\cr
Name of a currently selected attribute (if none, empty string).}
}
\value{
A \code{tibble} with a set of selected and possibly renamed columns.
}
\description{
Subsets columns to only use the ones relevant to the graph.
}
