% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prettify_chains.R
\name{sequence_indices}
\alias{sequence_indices}
\title{Index along a sequence}
\usage{
sequence_indices(sequence)
}
\arguments{
\item{sequence}{[\code{character(1)}]\cr
A single string with sequence to index.}
}
\value{
A vector of strings, each containing several padded indices to be
used as a single line.
}
\description{
Creates a vector of indices to use along with a sequence.
}
