% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-prettify_chains.R
\name{sequence_to_lines}
\alias{sequence_to_lines}
\title{Split sequence into lines}
\usage{
sequence_to_lines(sequence)
}
\arguments{
\item{sequence}{[\code{character(1)}]\cr
A single string with sequence to split into lines.}
}
\value{
A vector of strings, each element being its own line.
}
\description{
Splits sequence into lines of indicated length (by default 80)
and adds spaces between groups (by default 10 characters). To be used with
indices preferably.
}
