% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_handler.R
\name{table_download_handler}
\alias{table_download_handler}
\title{Handle download of AmyloGraph data}
\usage{
table_download_handler(input, table_data, write_function, extension)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the table.}

\item{table_data}{[\code{data.frame()}]\cr
AmyloGraph table data.}

\item{write_function}{[\code{function(1)}]\cr
A function that transforms table data to a file.}

\item{extension}{[\code{character(1)}]\cr
File extension without a dot to append to filename (e.g. \code{"csv"}).}
}
\value{
A \code{downloadHandler} object.
}
\description{
Creates a download handler that can be inserted into an app and
used to download AmyloGraph data in various formats.
}
