% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text_label_attribute}
\alias{text_label_attribute}
\title{Access human-readable attribute label}
\usage{
text_label_attribute(attribute)
}
\arguments{
\item{attribute}{[\code{character(1)}]\cr
An internal name of an interaction attribute.}
}
\value{
A single string with a human-readable label.
}
\description{
Returns a well-formatted label for an interaction attribute.
}
