% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text_question_attribute}
\alias{text_question_attribute}
\title{Access human-readable question text}
\usage{
text_question_attribute(attribute)
}
\arguments{
\item{attribute}{[\code{character(1)}]\cr
An internal name of an interaction attribute.}
}
\value{
A single string with a human-readable question.
}
\description{
Returns a well-formatted question asked when gathering data
about interactions.
}
