% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_row_selection.R
\name{toggle_state_and_css}
\alias{toggle_state_and_css}
\title{Enable/disable element with CSS styling}
\usage{
toggle_state_and_css(selector, class, condition)
}
\arguments{
\item{selector}{[\code{character(1)}]\cr
Query selector of the elements to target.}

\item{class}{[\code{character(1)}]\cr
Class to add when condition is not satisfied.}

\item{condition}{[\code{function(1)}]\cr
Function that takes no parameters and evaluates to either \code{TRUE} or \code{FALSE}.}
}
\description{
Toggles element state and sets CSS class to allow customizing
the look of elements when disabled (very useful to differentiate them from
enabled elements).
}
