% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{visAGNetwork}
\alias{visAGNetwork}
\title{Create visNetwork with AmyloGraph defaults}
\usage{
visAGNetwork(edges, nodes = ag_data_nodes)
}
\arguments{
\item{edges}{[\code{data.frame()}]\cr
AmyloGraph edge data.}
}
\value{
A \code{visNetwork} object that can be further modified.
}
\description{
Creates visNetwork with a set of defaults specific to
AmyloGraph.
}
