% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{visResetEdges}
\alias{visResetEdges}
\title{Clear and update edges in a graph}
\usage{
visResetEdges(graph, edges, input, selected_node_input_id)
}
\arguments{
\item{graph}{[\code{visNetwork_Proxy(1)}]\cr
Proxy of a graph to update.}

\item{edges}{[\code{data.frame()}]\cr
AmyloGraph graph data.}

\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the graph.}

\item{selected_node_input_id}{[\code{character(1)}]\cr
ID of the selected node (protein) to reselect.}
}
\value{
The same proxy as in \code{graph} parameter, but updated.
}
\description{
Clears all existing edges and creates new ones, then reselects
a node.
}
