% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{visToggleNodes}
\alias{visToggleNodes}
\title{Reselect a node in a graph}
\usage{
visToggleNodes(graph, id)
}
\arguments{
\item{graph}{[\code{visNetwork_Proxy(1)}]\cr
Proxy of a graph to update.}

\item{id}{[\code{character(0) | character(1)}]\cr
Node ID to verify.}
}
\value{
The same proxy as in \code{graph} parameter, but updated.
}
\description{
Reselects a node (clearing a previous selection, if exists) or
deselects all nodes, depending on value of \code{id}.
}
