% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_handler.R
\name{XGMML_download_handler}
\alias{XGMML_download_handler}
\title{Handle download of AmyloGraph data to XGMML}
\usage{
XGMML_download_handler(edges)
}
\arguments{
\item{edges}{[\code{reactivevalues()}]\cr
AmyloGraph data with "table" element.}
}
\value{
A \code{downloadHandler} object.
}
\description{
Creates a download handler that can be inserted into an app and
used to download AmyloGraph datain XGMML format. All proteins are used as
nodes, even is all its interactions are filtered out.
}
