% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{ag_data_proteins}
\alias{ag_data_proteins}
\title{AmyloGraph proteins list}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 78 rows and 4 columns.
}
\usage{
ag_data_proteins
}
\value{
\code{data.frame} coming from \code{protein_data.csv} file with the
following changes:
\itemize{
\item \code{id} contains unique hashes of protein names.
}
}
\description{
AmyloGraph proteins list
}
\keyword{datasets}
