% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGID_button.R
\name{AGID_button}
\alias{AGID_button}
\title{Generate AGID buttons}
\usage{
AGID_button(interaction_ids, id, ns, ...)
}
\arguments{
\item{interaction_ids}{[\code{character()}]\cr
AGIDs to use for buttons.}

\item{ns}{[\code{function(1)}]\cr
Namespace-generating function that takes single string as the only argument.
This is the namespace buttons should be placed in.}
}
\value{
A \code{character} vector the same length as \code{interaction_ids}, each
element being a string representation
}
\description{
Generates a vector of AGID buttons, where AGID codes are labels
of the buttons. Each button sets the \code{selected_interaction} value to the
respective AGID on click.
}
