% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_texts.R
\name{markdown_description}
\alias{markdown_description}
\alias{markdown_faq}
\alias{markdown_download}
\alias{markdown_citation}
\alias{markdown_contact}
\alias{markdown_acknowledgements}
\alias{markdown_yttutorial}
\alias{markdown_images}
\title{Access AmyloGraph section as a string}
\usage{
markdown_description()

markdown_faq()

markdown_download()

markdown_citation()

markdown_contact()

markdown_acknowledgements()

markdown_yttutorial()

markdown_images()
}
\value{
A single string.
}
\description{
This function retrieves a markdown string with a specific
section.
}
