% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_selected_node.R
\name{reactive_selected_node_label}
\alias{reactive_selected_node_label}
\title{Access label of currently selected protein}
\usage{
reactive_selected_node_label(input)
}
\arguments{
\item{input}{[\code{reactivevalues()}]\cr
An input object of the app or a module that contains the protein selector.}

\item{node_data}{[\code{data.frame()}]\cr
AmyloGraph node data.}
}
\value{
A \code{reactive} object with a single string containing protein
name.
}
\description{
This function extracts label of the protein selected in one of
the available selectors (drop-down list or clicking on graph node).
}
