% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_network.R
\name{render_network}
\alias{render_network}
\title{Render visNetwork graph}
\usage{
render_network(ag_data_nodes, edges)
}
\arguments{
\item{ag_data_nodes}{[\code{data.frame()}]\cr
Output of \code{\link{ag_data_nodes}()}, nodes to plot.}

\item{edges}{[\code{reactivevalues()}]\cr
AmyloGraph data with "graph" element, edges to plot.}
}
\description{
Renders a visNetwork graph using initial AmyloGraph data. Any
changes to the graph should be done through \code{visNetworkProxy()} inside an
observer in other parts of code.
}
